--- src/modules/render-films-number.js
@@ -1,8 +1,0 @@
-import NumberOfFilms from '../view/number-of-films';
-import {renderDOMElement, Positions} from '../utils/render';
-
-export const renderFilmsNumber = (data) => {
-  const numberOfFilmsContainer = document.querySelector('.footer__statistics');
-  const numberOfFilms = new NumberOfFilms(data);
-  renderDOMElement(numberOfFilmsContainer, numberOfFilms, Positions.BEFOREEND);
-};
