--- src/modules/render-filters-menu.js
@@ -1,8 +1,0 @@
-import FiltersMenu from '../view/filters-menu';
-import {renderDOMElement, Positions} from '../utils/render';
-
-export const renderFiltersMenu = (data) => {
-  const mainContainer = document.querySelector('.main');
-  const filtersMenu = new FiltersMenu(data);
-  renderDOMElement(mainContainer, filtersMenu, Positions.AFTERBEGIN);
-};
