+++ src/presenter/films-list-presenter.js
@@ -11,7 +11,6 @@ import {showPopup} from '../modules/show-popup';
 import MoreButton from '../view/more-button';
 import {showMoreCards} from '../modules/show-more-cards';
 import FilmCard from '../view/film-card';
-import {renderExtraData} from '../modules/render-extra-data';
 import {sortByCommentsNumber, sortByRating} from '../modules/data-sort';
 import {getAllMovies} from '../modules/data-filters';
 
@@ -107,13 +106,22 @@ export default class FilmsListPresenter {
     }
   }
 
+  _renderExtraData(container, data) {
+    const NUMBER_OF_EXTRA_CARDS = 2;
+    const numberToShow = Math.min(data.length, NUMBER_OF_EXTRA_CARDS);
+    for (let i = 0; i < numberToShow; i++) {
+      const filmCard = new FilmCard(data[i]);
+      renderDOMElement(container, filmCard, Positions.BEFOREEND);
+    }
+  }
+
   _renderTopRated(data) {
     const topRatedContainer = document.querySelectorAll('.films-list__container')[1];
-    renderExtraData(topRatedContainer, sortByRating(data));
+    this._renderExtraData(topRatedContainer, sortByRating(data));
   }
 
   _renderMostCommented(data) {
     const mostCommentedContainer = document.querySelectorAll('.films-list__container')[2];
-    renderExtraData(mostCommentedContainer, sortByCommentsNumber(data));
+    this._renderExtraData(mostCommentedContainer, sortByCommentsNumber(data));
   }
 }
